feat: add library with warnings and examples
